+++ common/src/main/java/com/fej1fun/potentials/fluid/UniversalFluidStorage.java
@@ -9,5 +9,5 @@ public interface UniversalFluidStorage extends Iterable<FluidStack> {
     boolean isFluidValid(int tank, FluidStack stack);
     long fill(FluidStack stack, boolean simulate);
     FluidStack drain(FluidStack stack, boolean simulate);
-    FluidStack drain(int maxAmount, boolean simulate);
+    FluidStack drain(long maxAmount, boolean simulate);
 }
